+++ packages/token-bridge-sdk/src/hooks/arbTokenBridge.types.ts
@@ -181,12 +181,9 @@ export interface ArbTokenBridgeEth {
 }
 
 export interface ArbTokenBridgeToken {
-  tokensFromLists: ContractStorage<ERC20BridgeToken>
-  tokensFromUser: ContractStorage<ERC20BridgeToken>
   add: (erc20L1orL2Address: string) => Promise<void>
   addTokensFromList: (tokenList: TokenList, listID: number) => void
   removeTokensFromList: (listID: number) => void
-  searchTokenFromList: (erc20L1orL2Address: string) => ERC20BridgeToken | null
   updateTokenData: (l1Address: string) => Promise<void>
   approve: (params: {
     erc20L1Address: string
