+++ packages/token-bridge-sdk/src/hooks/useTokenLists.ts
@@ -1,9 +1,13 @@
+/*
+  Hook and Utility functions meant for fetching and maintaining the token-lists on the basis of our existing hardcoded `BRIDGE_TOKEN_LISTS`
+  - will have SWR caching
+*/
+
 import { useEffect, useState } from 'react'
 import axios from 'axios'
 import { TokenList } from '@uniswap/token-lists'
 import {
   ArbTokenBridge,
-  ContractStorage,
   ERC20BridgeToken,
   L1TokenData,
   TokenType
@@ -182,18 +186,6 @@ export function fetchTokenLists(): Promise<void> {
   })
 }
 
-export function useTokenLists(forL2ChainId?: string): TokenListWithId[] {
-  const [tokenLists, setTokenLists] = useState<TokenListWithId[]>(() =>
-    getTokenLists(forL2ChainId)
-  )
-
-  useEffect(() => {
-    setTokenLists(getTokenLists(forL2ChainId))
-  }, [forL2ChainId])
-
-  return tokenLists
-}
-
 export function getTokenLists(forL2ChainId?: string): TokenListWithId[] {
   const storage = sessionStorage.getItem(STORAGE_KEY)
 
@@ -210,88 +202,6 @@ export function getTokenLists(forL2ChainId?: string): TokenListWithId[] {
   return parsedStorage.filter(tokenList => tokenList.l2ChainId === forL2ChainId)
 }
 
-export function tokenListsToSearchableTokenStorage(
-  tokenLists: TokenListWithId[],
-  l1ChainId: string,
-  l2ChainId: string
-): ContractStorage<ERC20BridgeToken> {
-  return tokenLists.reduce(
-    (acc: ContractStorage<ERC20BridgeToken>, tokenList: TokenListWithId) => {
-      tokenList.tokens.forEach(token => {
-        const address = token.address.toLowerCase()
-        const stringifiedChainId = String(token.chainId)
-        const accAddress = acc[address]
-
-        if (stringifiedChainId === l1ChainId) {
-          // The address is from an L1 token
-          if (typeof accAddress === 'undefined') {
-            // First time encountering the token through its L1 address
-            acc[address] = {
-              ...token,
-              type: TokenType.ERC20,
-              l2Address: undefined,
-              listIds: new Set()
-            }
-          } else {
-            // Token was already added to the map through its L2 token
-            acc[address] = {
-              ...accAddress,
-              address
-            }
-          }
-
-          // acc[address] was defined in the if/else above
-          acc[address]!.listIds.add(tokenList.bridgeTokenListId)
-        } else if (stringifiedChainId === l2ChainId) {
-          // The token is an L2 token
-
-          if (!token.extensions?.bridgeInfo) {
-            return
-          }
-
-          // @ts-ignore TODO
-          // TODO: should we upgrade '@uniswap/token-lists'?
-          const bridgeInfo: {
-            [chainId: string]: { tokenAddress: string }
-          } = token.extensions.bridgeInfo
-
-          const l1Bridge = bridgeInfo[l1ChainId]
-          if (l1Bridge) {
-            const addressOnL1 = l1Bridge.tokenAddress.toLowerCase()
-
-            if (!addressOnL1) {
-              return
-            }
-
-            if (typeof acc[addressOnL1] === 'undefined') {
-              // Token is not on the list yet
-              acc[addressOnL1] = {
-                name: token.name,
-                symbol: token.symbol,
-                type: TokenType.ERC20,
-                logoURI: token.logoURI,
-                address: addressOnL1,
-                l2Address: address,
-                decimals: token.decimals,
-                listIds: new Set()
-              }
-            } else {
-              // The token's L1 address is already on the list, just fill in its L2 address
-              acc[addressOnL1]!.l2Address = address
-            }
-
-            // acc[address] was defined in the if/else above
-            acc[addressOnL1]!.listIds.add(tokenList.bridgeTokenListId)
-          }
-        }
-      })
-
-      return acc
-    },
-    {}
-  )
-}
-
 export function toERC20BridgeToken(data: L1TokenData): ERC20BridgeToken {
   return {
     name: data.name,
@@ -301,4 +211,16 @@ export function toERC20BridgeToken(data: L1TokenData): ERC20BridgeToken {
     decimals: data.decimals,
     listIds: new Set()
   }
+}
+
+export function useTokenLists(forL2ChainId?: string): TokenListWithId[] {
+  const [tokenLists, setTokenLists] = useState<TokenListWithId[]>(() =>
+    getTokenLists(forL2ChainId)
+  )
+
+  useEffect(() => {
+    setTokenLists(getTokenLists(forL2ChainId))
+  }, [forL2ChainId])
+
+  return tokenLists
 }
